Reorganize names and file structure
This commit takes the merged files of the prior commit and
reorganizes the files and names to make this a suitable base
for the term project and exercises in Spring 2021 and beyond.

Files:
1. Merge the e-obs directory into e-k8s
2. Merge the misc subdirectory into cluster
3. Centralize all the templated values into `cluster/tpl-values.txt`
4. Rename any files that require template processing as `*-tpl.*`.

Names:
1. Name the release `c756`
2. Name the application namespace `c756ns`
3. Name the application gateway `c765-gateway`
4. Name the application virtual service `c756vs`

See `cluster/tpl-values.txt` for a list of the template variables.
Note that the AWS security labels are in this file and will be
inserted into `cluster/awscred-tpl.yaml` when the template processor
is run.

Note also that this commit DOES NOT INCLUDE an actual template
processor. That will be added in a later commit.

Possible future improvements:
1. The names might all themselves be templated. Perhaps make everything Helm charts?
2. The name `c756vs` is only one letter different from `c756ns`.  Maybe make it more distinct?