Do not trace health checks
Turn off Prometheus tracing of liveness and readiness checks.

Also bring code into compliance with `flake8` coding style. This
checks for PEP-8 formatting and a reasonable set of style
requirements.

One flake8 warning left in: Retaining assignments to `header` variable
that are never used.  This makes it clearer for students who may wish
to examine headers in their modifications to this code.