Add target to provision entire system
The target `provision` will attempt to execute most of the lines specified
in `README.md`, excluding the one-time ones such as adding entries to
the helm search path.  It is intended to be run when you bring up a
new cluster on a cloud provider.

The current implementation is brittle and considered beta.  It will
need to be tested and made more robust but it is a start.

The implementation uses shell scripts `tools/waiteq.sh` and
`tools/waitne.sh` to wait for long-running installes to complete
before procedding to the next step.  These are likely the most brittle
parts of the sequence.

Possible extensions:

Currently, the two `wait*.sh` scripts presume that the pod waited upon
is within the `istio-system` namespace.  A reasonable extension would
be to add yet another argument, specifying that.