Add ServiceEntry for DynamoDB
This identifies a DynamoDB endpoint to Istio. The endpoint will
appear, labelled as "dynamodb-aws", on the Kiali graph.

This version of the file is templated for the region in which DynamoDB
is running. It presumes the endpoint port is 1440.  I am unsure whether AWS is
committed to this port (either generally, or just for DynamoDB).  If
other instances of DynamoDB use a different port, the port value in
this file will have to be updated.