Plot total requests / service (incl. errors)
The previous version of this dashboard plotted separate series for
each return code.  This commit changes the query to report totals by
service for all return codes.

Viewers can determine error rates from the error chart immediately
below this one.

This change also changes the series label from `job` to
`service`. This will put the same labels on the chart but `service` is
used in other plots on this dashboard and for future consistency this
plot should agree with them.