Add a cmpt756db VirtualService
All our application services should also be VirtualServices (VS), so
that they are known to the service mesh.

The default file `cluster/database-vs.yaml` is a straightforward
VS. The two other manifests, `database-vs-delay.yaml` and
`database-vs-fault.yaml`, insert delays and 500 faults respectively,
for a fraction of the requests.  These manifests will be used in
Exercise 6 to demonstrate inducing faults using Istio.