Standardize vendor *.mak files
Standardize on a single small set of targets for all the vendor *.mak
files (where Minikube is also considered a "vendor"):

start: create a fresh cluster
stop: stop and delete the cluster
ls: list any clusters for this vendor
status: same as ls but tee output into a log
cd: switch the (presumed unique) context for this vendor
up: bring up a set of worker nodes (AWS only; prints message on
    others).  Must be followed by "make -f k8s.mak reinstate"
down: tear down a set of worker nodes (AWS only; prints message on others)

Along the way, standardized on all vendor files having a `*-tpl.mak`
version that is the canonical representation, whether or not the
current version of the file uses any template variables.  This also
meant that `az.mak` and `mk.mak` needed to be removed from the repo
and their names placed in `.gitignore`.  Only their template versions
should be updated and recorded.