Delete Kubernetes context after deleting cluster
This commit ensures that the list of Kubernetes contexts matches the
list of currently active clusters.  Before this, the context for a
deleted cluster remained, creating the confusing appearence that those
clusters were still running when in fact they were gone but the local
list on the client retained their name.

A cluster must be completely deleted before deleting its associated
context or else `kubectl` will give a warning.  To ensure complete
deletion, the delete calls for Azure and GCP had to change from
asynchronous to synchronous.

To delete a context without getting a Kubernetes warning, the cluster must