Use t3.medium nodes in AWS
Creating clusters of `t2.medium` nodes fails frequently on AWS.  As
the number of machines supporting such nodes is decreasing in the AWS
Availability Zones, this commit upgrades the requested machines to
`t3.medium`, which are more widely supported.

This increases the charges (perhaps even preventing these clusters
from running in some account types?) but there was no point in
continuing to request unavailable machine types.