Add Version 2 to S2
This substantial commit makes all the changes necessary to add a
section version to S2.  The S1 and DB services remain unchanged, with
only a single version. Including both styles in this repo provides
examples of each approach.

Adding a second version requires

1. Adding a new directory layer to S2 for the versions.
2. Separating the VirtualService, Deployment, and Gateway definitions for all
   services.
3. Adding a DestinationRule for the S2 VirtualService.
4. Defining a DestinationRule/VirtualService pair for each version of
   S2.

This commit does some minor housekeeping to the definitions in
subdirectory cluster:

1. Adding comments
2. Changing the container names so that they match the service definitions.