Add loader to initialize the database
This update adds a Kubernetes Job to initialize the database by adding
the users in the Gatling resource `users.csv` and the songs in the
Gatling resource `music.csv`.

The loader is located in subdirectory `/load`.

The database code is extended by a `/load` route.  This differs from
`/write` in that it is passed a UUID rather than generates a new
id. This route also requires an `Authorize` header in Basic format
for userid `svc-loader` and password equal to the value of Secret
`svc-loader-token`.  The loader also accesses that Secret so that it
can be passed in the header.

A new tool is added, `tools/build-configmap.sh` to construct a
ConfigMap from a text file.  The header files
`cluster/music-header.yaml` and `cluster/users-header.yaml` define the
YAML header that must precede the data.

The cluster makefile `kis-tpl.mak` is updated and reorganized to
supporst the needs of the loader.