Rename target 'gatling'
The original target name, `gatling`, inadvertently matched an actual
directory name. The target would only "work" when that directory had
an older update time than file
`gatling/simulations/proj756/ReadTableSim.scala`.

This revision renames the target to a name that does not correspond to
any filename. This ensures the target will always be built when made,
which is its intent.