Add optional image version argument to shell.sh
This isn't the friendliest interface---it doesn't check for an
argument of -h or otherwise display a usage message---but it gets
the job done for the moment.

Ultimately, production versions will have a multi-architecture manifest,
pulling the correct image for the user's architecture. This is just a fix
to permit developing the images for all the architectures.